+++ src/api/v1/urls.py
@@ -16,6 +16,7 @@ urlpatterns = [
     re_path(r"^accounts/", include("api.v1.accounts.urls")),
     re_path(r"^app/", include("api.v1.app.urls")),
     re_path(r"^blog/", include("api.v1.blog.urls")),
+    re_path(r"^collections/", include("api.v1.collections.urls")),
     re_path(r"^events/", include("api.v1.events.urls")),
     re_path(r"^home/", include("api.v1.home.urls")),
     re_path(r"^invites/", include("api.v1.invites.urls")),