+++ src/collection/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+(C) 2013-2025 Copycat Software, LLC. All Rights Reserved.
+"""
+
+# pylint: disable=wildcard-import
+# pylint: disable=unused-wildcard-import
+from collection.models.Collection import *